[PATCH] uml: Fix the console stuttering
I spent far too much of the weekend tracking this sucker down through the guts
of the tty code.  The problem turns out to be that drivers/char/n_tty.c has a
write_chan that does buffering and retransmitting data, and
arch/um/drivers/chan_kern.c ALSO has a write_chan that buffers and retransmits
data, and the first calls the second but the second doesn't always return
correct status information for the -EAGAIN case.

When they get confused, both of them try to buffer and retransmit data, hence
the stuttering.

The first fix is that if chan_kern's write_chan gets an -EAGAIN, it should NOT
gratuitously change that to a 0 before returning.  I don't know why that code
is in there, but deleting those two lines makes 90% the stuttering go away.
But not quite all of it.

The second half of the fix is arch/um/drivers/line.c has a buffer_data()
function that adds data to the buffer, tries to flush the buffer out to disk,
gets -EAGAIN, and then returns -EAGAIN even though it successfully buffered
all the data it was sent.  So the upper layer resubmits the last chunk of data
it sent when the console unblocks, even though the lower layer buffered it and
sent it on by that point.

With this patch, I can't get the UML console to stutter anymore by suspending
the process it's writing to.  (Add tee to the mix and you can still make it
hang by suspending its xterm for a second or two, but I think that tee is
hanging, not UML.  Hangs with RHEL4 tee, but not busybox tee...)

Signed-off-by: Rob Landley <rob@landley.net>
Acked-by: Jeff Dike <jdike@addtoit.com>
Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade@yahoo.it>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 424c271e-8QJo7TeVPRbbaiCB9ocAw