[PATCH] paport oops fix
It appears that the parport driver claims on-board superio devices without
actually doing anything.  When the driver is removed, we try to dereference
non-existent driver data to unregister the ports.  Since we didn't register
anything, it's safe to ignore these devices in the remove function.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 424c27ce56j4xGLTw7yNs_iIKbqXZw