[PATCH] fbdev: mvidia licensing clarification
Mark said that the nVidia licensing is MIT-style, so it should be okay.
I'm including a patch that adds a small licensing comment on the pertinent
files and also adds the license notice in nv_accel.c, just to be on the
safe side.

Signed-off-by: Antonino Daplas <adaplas@pol.net>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 424c2877dt8c0B_HSxcjSWW7xZ8vCg