[PATCH] nvidiafb: Maximize blit buffer capacity
Currently, nvidiafb blit expands a rectangle a scanline at a time.  This is
very inefficient because the graphics pipeline is not maximized.  It may also
crash if the scanline is > 4096 (should not happen with fbcon though).

Instead of blitting a scanline at a time, attempt to fill up buffer to maximum
capacity.

Signed-off-by: Antonino Daplas <adaplas@pol.net>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 424c2899JfG6_2K1XpviZ_SR89hSuA