[PATCH] radeonfb: Preserve TMDS setting
radeonfb tends to override TMDS_TRANSMITTER_CNTL completely, which appears
to be incorrect with some recent flat panels.  This patch changes this so
that we only tweak the bits we want in there and preserve the BIOS value
for the rest.

Along with the i2c fix, that patch fixes operations on the latest Apple
cinema displays.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 424c2a18CWm_0pSIh7CAvGaC4UcyIQ