[PATCH] nfsd4_remove_unused_acl_function
We're translating v4 ACL's to POSIX ACL's and relying on the filesystem to
check those, so there's no need for a NFSv4 ACL-checking algorithm.

Some day we'd like to find a way to implement the full v4 ACL protocol on the
server side, but for now this is dead code.

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>
Signed-off-by: Neil Brown <neilb@cse.unsw.edu.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 424c2c83uygYj6mxI98aIyYC4tBzuQ